File serving improvements, 404 error handling
WHAT: File serving now reads the request target to determine the file to serve rather than using a hardcoded value. Relies on the dispatcher to substitute the relative path (notes on this below). The relative path substitution is currently hardcoded.
WHY: This allows serving different files instead of just the home page.
TESTING: Unit and integration tests pass. Behavior is changed when requesting a file using local browser, but this was expected. Previously, the request for style.css would simply return index.html again, but now the browser receives a 404 error for style.css (which has not been created yet). The browser still correctly receives index.html from the main request.

WHAT: Implemented 404 error handling. If the requested file could not be found, serve the 404 page (located at files_to_serve/404.html).
WHY: This is part of what I would consider "base" functionality for a web server.
TESTING: Old unit and integration tests pass. Local browser is able to receive the 404 error code and the intended 404 page. New tests for this feature pending.

WHAT: Updated unit test HTML input files to remove links to style.css.
WHY: The old unit test was failing due to receiving a 404 error from the CSS file sub-request. Now, only one request is made per unit test.
TESTING: Unit and integration tests pass.

Notes on performing relative path substitution in the dispatcher:
Relative path substitution refers to mapping a request's target URI to the actual relative file path on the server; for example, for a request "GET / HTTP/1.1" with target URI "/", the relative file path to substitute is the home page, "/files_to_serve/index.html".

Performing relative path substitution in the dispatcher rather than in the FileRequestHandler is admittedly not optimal design practice because the FileRequestHandler is no longer implementation-agnostic. However, I think the trade-off is worth it in this case. The dispatcher's responsibility is to determine which type of RequestHandler to dispatch. It does this by reading the prefix of the request target and matching it to the prefixes defined in the server's config file. The part of the request target following the prefix is the relative path of the file. Thus, if the dispatcher is going to dispatch a FileRequestHandler, it already knows the relative file path, and it is convenient to substitute the relative file path in the request's target field before passing it to the FileRequestHandler. If the FileRequestHandler were to do this process (such that it becomes implementation-agnostic), the matching process (which is computationally expensive) would have to be repeated. An alternative would be to pass the relative path as an argument to the FileRequestHandler's constructor, but this is less memory efficient and would also complicate the API design; because all RequestHandlers share a common interface, an additional string would have to be passed to all RequestHandlers.