Dispatching, relative path substitution
WHAT: The server now dynamically dispatches the correct RequestHandler type based on the HTTP request's target and the URI mapping. Config parsing is not yet implemented, so the URI mapping is currently hardcoded.
WHY: The server should be able to support different types of RequestHandlers based on URIs, and it should be able to do so in a configurable manner.
TESTING: Unit test and integration test pass. Trace logs show correct dispatching behavior. Behavior is not changed when requesting a file using local browser.

WHAT: The dispatcher performs relative path substitution on the HTTP request's target when dispatching a FileRequestHandler.
WHY: It saves a significant amount of computation compared to doing so in the FileRequestHandler itself. See previous commit for a more detailed explanation.
TESTING: Unit test and integration test pass. Behavior is not changed when requesting a file using local browser.