Trace logs suppressed by default
WHAT: Trace logs are now suppressed by default and must be enabled using Log::enable_trace(). (Eventually this should be configurable)
WHY: The next feature to be implemented will generate a lot of trace logs, so it is nice to be able to turn trace logs on and off.
TESTING: log_test updated to test both enabled and suppressed trace logs. Unit and integration tests pass. Feature works as expected.