Config parsing framework
WHAT: Implemented a framework for config parsing.
- src/server_main.cc invokes the config parser. Updated usage string.
- src/nginx_config_parser.cc is able to parse tokens from config and generate trace logs. Does not actually process meaningful information from the config or validate the structure yet.
- run.sh is updated to use a config file instead of a port
- Created configs folder with configs/local_config.conf
- Tiny update to README.md documentation (just looks better, previous version was still correct)
WHY: It is important for a web server to be externally configurable, and this is a core feature listed in the design documents.
TESTING: New tests for config parsing feature pending feature completion. Old unit and integration tests pass. Trace logs are as expected when parsing configs/local_config.conf.