Modified config parser to singleton class
WHAT: Modified the implementation of the NGINX config parser to use a singleton class instead of instantiating a parser object in main. Renamed Parser to Config because it makes more sense in this context.
WHY: We only need one config parser, so a singleton makes sense. Furthermore, this simplies the API of server, session, and RequestHandler since we no longer need to pass around multiple copies of the same information as arguments.
TESTING: Behavior is unchanged. Unit and integration tests pass.