Set up React frontend with navigation bar, styling
WHAT: Set up React frontend with a barebones website that contains a navigation bar and styling.
WHY: This is going to be the framework of the website.
TESTING: Homepage serves correctly. Other pages do not serve correctly, planning to split the fix into the next commit. Unit tests pass, integration test expectedly fails due to the change of structure.