Added WIP web interface for project binaries
WHAT: Added web interface that allows the user to run the binary for a project. The server doesn't run any binary yet, just gets the input data from the web form and returns a placeholder response.
WHY: I think this is a cool feature to have.
TESTING: Server is able to read a POST request sent from the form and return a placeholder response.

Might have to consider rate limiting this feature before public release to limit server load.