Updates to READMEs
WHAT: Updated README.md to indicate that the back end build script 'build.sh' does not build the front end, and that errors may occur in testing if the front end has not been built.
WHY: It may be confusing if a user follows the build instructions and gets errors.

WHAT: Fixed an error in the title of frontend/README.md.