Improved logging of invalid requests
WHAT: Invalid requests are now logged in their entirety.
WHY: I would like to see the contents of the invalid requests I am getting, in particular to see whether or not it is a bug with my server or simply an invalid request.
TESTING: Added new integration test for request with invalid method. Unit and integration tests pass.