Added featured projects carousel to homepage
WHAT: Added a carousel to the homepage that features selected projects.
WHY: Maybe not every visitor will care enough to go to the projects page if the homepage is too boring, so I wanted to add something eye catching to the homepage to entice visitors to look at some projects.
TESTING: Carousel displays as desired. Unit and integration tests pass.