Split server and session into HTTP and HTTPS
WHAT: Implemented HTTP and HTTPS versions of `server` and `session` that can co-exist.
WHY: This is important for implementing HTTP redirect to HTTPS, as the server must be able to accept both types of requests simultaneously.
TESTING: Unit and integration tests pass. Server correctly serves either HTTPS content or HTTP content depending on the config, but not both (see below).

TODO: There is a segmentation fault bug that occurs when a HTTPS server is on port 8080 and the HTTP server is on port 8081 (I am unsure if the port order matters). Will have to investigate this another time.
TODO: See if it's possible to externalize the lambda write handler in `session.cc` to have less to override.
TODO: Consider separate classes for HTTP and redirect.
TODO: Implement SSL features in `nginx_config_parser`.